% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-collect_rxclass_context.R
\name{collect_rxclass_context}
\alias{collect_rxclass_context}
\title{Collect RxClass Context Data}
\usage{
collect_rxclass_context(
  class_types = c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISEASE", "DISPOS",
    "CHEM", "SCHEDULE", "STRUCT")
)
}
\arguments{
\item{class_types}{Vector of desired classTypes. This vector is also
in the order of the API calls will be made. Can be one or more of the following:
"MESHPA", "EPC", "MOA", "PE", "PK", "DISEASE", "DISPOS", "CHEM",
"SCHEDULE", "STRUCT", "TC", "VA", "ATC1-4".}
}
\description{
Get the ancestors of a class specified by source and classId. The ancestors are returned as a graph of nodes and edges.The source, which restricts the results to a particular class type, should be specified when using a MeSH identifier because some MeSH concepts exist in more than one RxClass tree. getClassGraphBySource returns only one tree. To find all trees that contain a concept, use getClassContexts (\verb{*_rxclass_context}).This resource returns an array of nodes (rxclassMinConceptItem) and an array of edges (rxclassEdge) that link the nodes.
}
