% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_rxclass_graph.R
\name{collect_rxclass_graph}
\alias{collect_rxclass_graph}
\title{Collect RxClass Graph Data}
\usage{
collect_rxclass_graph(
  class_types = c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISEASE", "DISPOS",
    "CHEM", "SCHEDULE", "STRUCT"),
  prior_version = NULL,
  prior_api_version = "3.1.174"
)
}
\arguments{
\item{class_types}{Vector of desired classTypes. This vector is also
in the order of the API calls will be made. Can be one or more of the following:
"MESHPA", "EPC", "MOA", "PE", "PK", "DISEASE", "DISPOS", "CHEM",
"SCHEDULE", "STRUCT", "TC", "VA", "ATC1-4".}
}
\description{
Collect RxClass Graph Data
}
\seealso{
Other Collect functions: 
\code{\link{collect_rxclass_members}()}
}
\concept{Collect functions}
\concept{RxClass Graph functions}
