% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_rxclass_data.R
\name{dev_rxclass_data}
\alias{dev_rxclass_data}
\title{FUNCTION_TITLE}
\usage{
dev_rxclass_data(
  rela_sources = c("DAILYMED", "MESH", "FDASPL", "FMTSME", "VA", "MEDRT", "RXNORM",
    "SNOMEDCT"),
  class_types = c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISPOS", "SCHEDULE",
    "STRUCT")
)
}
\arguments{
\item{rela_sources}{PARAM_DESCRIPTION, Default: c("DAILYMED", "MESH", "FDASPL", "FMTSME", "VA", "MEDRT", "RXNORM",
"SNOMEDCT")}

\item{class_types}{PARAM_DESCRIPTION, Default: c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISPOS", "SCHEDULE",
"STRUCT")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
