% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rxclass_graph.R
\name{extract_rxclass_graph}
\alias{extract_rxclass_graph}
\title{Extract RxClass Graph}
\usage{
extract_rxclass_graph(
  class_types = c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISEASE", "CHEM",
    "SCHEDULE", "STRUCT", "DISPOS")
)
}
\description{
Extract RxClass Graph
}
\details{
Some class types such as DISEASE follow a cyclical hierarchical pattern that
can recurse into infinity if a limit is imposed when writing the Concept Ancestor
csv file. Concept Ancestor processing is completed when a subsequent level of
parent and children is a dataframe that is a duplicate of a previous level's.
}
