% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rxclass_members.R
\name{extract_rxclass_members}
\alias{extract_rxclass_members}
\title{Extract RxClass Members}
\usage{
extract_rxclass_members(
  rela_sources = c("DAILYMED", "MESH", "FDASPL", "FMTSME", "VA", "MEDRT", "RXNORM",
    "SNOMEDCT"),
  class_types = c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISEASE", "DISPOS",
    "CHEM", "SCHEDULE", "STRUCT"),
  prior_version = NULL,
  prior_api_version = "3.1.174"
)
}
\description{
Extract RxClass Members
}
\details{
RxClass Members represent the relationship between a Class belonging to
RxClass Graph and a Drug concept such as an RxNorm Ingredient.
}
