% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{get_rxnorm_version}
\alias{get_rxnorm_version}
\title{Get RxNorm Version}
\usage{
get_rxnorm_version(
  conn,
  conn_fun = "pg13::local_connect()",
  checks = c("conn_status", "conn_type", "rows", "source_rows"),
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\description{
Get RxNorm Version
}
