% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_rxnorm_archive_update.R
\name{process_rxnorm_concept_update}
\alias{process_rxnorm_concept_update}
\title{Process the RxNorm Concept Update Table}
\usage{
process_rxnorm_concept_update(
  conn,
  conn_fun = "pg13::local_connect()",
  checks = "",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\description{
Process the RxNorm Concept Update Table
}
