% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rxnorm_validity_status.R
\name{process_rxnorm_validity_status}
\alias{process_rxnorm_validity_status}
\title{Process the RxNorm Validity Table}
\usage{
process_rxnorm_validity_status(
  conn,
  conn_fun = "pg13::local_connect(verbose = {verbose})",
  processing_schema = "process_rxnorm_validity_status",
  destination_schema = "rxtra",
  rm_processing_schema = TRUE,
  checks = "",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE,
  prior_version = NULL,
  prior_api_version = "3.1.174"
)
}
\description{
Retrieve and cache concept batches by
its validity status from
the RxNav REST API and write a
table that maps an input RxCUI or RxNorm
code to its most current RxCUI or RxNorm code.
}
\details{
Default Schema: \code{rxtra}

Time Requirement: Approximately 30 minutes to an hour

Resource Requirements:
\itemize{
\item RxNorm schema
\item RxNorm API
}
}
