% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{requires_processing}
\alias{requires_processing}
\alias{log_processing}
\title{Check If A Table Requires Processing}
\usage{
requires_processing(
  conn,
  conn_fun = "pg13::local_connect()",
  checks = "",
  target_table,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)

log_processing(
  conn,
  conn_fun = "pg13::local_connect()",
  checks = "",
  target_table,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\description{
Check If A Table Requires Processing

Log Processing of a Table
}
