% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupRxNorm-package.R
\docType{package}
\name{setupRxNorm}
\alias{setupRxNorm}
\alias{setupRxNorm-package}
\title{setupRxNorm: A package for RxNorm and RxClass data management}
\description{
Instantiate NIH's RxNorm monthly release and process RxNav RxClass API calls in relational database format.
}
\author{
\strong{Maintainer}: Meera Patel, MD \email{patelmeeray@gmail.com} (\href{https://orcid.org/0000-0002-3401-3948}{ORCID})

}
\keyword{internal}
