% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_rxclass_members.R
\name{collect_rxclass_members}
\alias{collect_rxclass_members}
\title{Collect RxClass Members Data}
\usage{
collect_rxclass_members(
  rela_sources = c("DAILYMED", "MESH", "FDASPL", "FMTSME", "VA", "MEDRT", "RXNORM",
    "SNOMEDCT"),
  class_types = c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISEASE", "DISPOS",
    "CHEM", "SCHEDULE", "STRUCT"),
  prior_version = NULL,
  prior_api_version = "3.1.174"
)
}
\arguments{
\item{rela_sources}{Vector of desired classTypes. This vector is also
in the order of the API calls will be made. Can be one or more of the following:
'DAILYMED',
'MESH',
'FDASPL',
'FMTSME',
'VA',
'MEDRT',
'RXNORM',
'SNOMEDCT'.}
}
\description{
Collect RxClass Members Data
}
\seealso{
Other Collect functions: 
\code{\link{collect_rxclass_graph}()}
}
\concept{Collect functions}
