% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rxclass_members.R
\name{extract_rxclass_members}
\alias{extract_rxclass_members}
\title{Extract RxClass Members}
\usage{
extract_rxclass_members(
  rela_sources = c("DAILYMED", "MESH", "FDASPL", "FMTSME", "VA", "MEDRT", "RXNORM",
    "SNOMEDCT"),
  class_types = c("MESHPA", "EPC", "MOA", "PE", "PK", "TC", "VA", "DISEASE", "DISPOS",
    "CHEM", "SCHEDULE", "STRUCT")
)
}
\description{
Extract RxClass Members
}
